+++ changelog.md
@@ -1,4 +1,15 @@
 # Changelog
 
 **./app/test-common.js**
-* Merged 'checkPropertyStringRequiredObject' into 'checkPropertyStringRequiredArray'
+* Moved 'checkPropertyStringRequiredArray' function to 'test-array.js' as 'checkAllStringRequired'
+* Commented out 'testPropertyStringRequiredArray' export.
+
+---
+
+**Affected Files**
+* ./test-parts/
+	* part-h-api_requests/requests/
+		* req-admin_main.js
+		* req-alarm.js
+	* part-i-api_frontend/parts/
+		* g-get_device_properties.js